Add files via upload
Problem Statement:
1) Exploratory Data Analysis (EDA):
Analyze the distribution of stock prices and volumes for each company over time.
Visualize trends, seasonality, and patterns in the stock market data using line charts, bar plots, and heatmaps.

2)Correlation Analysis:
Investigate the correlations between the closing prices of different companies to identify potential relationships.
Calculate correlation coefficients and visualize correlation matrices.

3)Top Performers Identification:
Identify the top-performing companies based on their stock price growth and trading volumes over a specific time period.

4)Market Sentiment Analysis:
Perform sentiment analysis using Natural Language Processing (NLP) techniques on news headlines related to each company.
Determine whether positive or negative news impacts the stock prices and volumes.

Column Info:
Company : The stock ticker symbol of the company, used to uniquely identify it in the stock market
Date : The date on which the stock market data was recorded or reported. It represents the trading day for which the stock
close/last : The closing price or the last traded price of the company's stock on the given date. It represents the final price at which
Volume : The total number of shares of the company's stock traded on the given date. It indicates the level of investor interest 
Open : The opening price of the company's stock on the given date. It is the price at which the first trade occurred during the 
High : The highest price at which the company's stock traded on the given date. It indicates the highest price reached 
Low : The lowest price at which the company's stock traded on the given date. It indicates the lowest price reached